init simple, generic, version of calculating vertex attributes

+++ DX12Lib/inc/dx12lib/CommandList.h
@@ -627,7 +627,7 @@ protected:
 private:
     // Used for procedural mesh generation.
     using VertexCollection = std::vector<dx12lib::VertexPositionNormalTangentBitangentTexture>;
-    using IndexCollection  = std::vector<uint16_t>;
+    using IndexCollection  = std::vector<uint32_t>;
 
     // Create a scene that contains a single node with a single mesh.
     std::shared_ptr<Scene> CreateScene( const VertexCollection& vertices, const IndexCollection& indicies );
